+++ core/src/edge.cpp
@@ -4,7 +4,7 @@
 
 using namespace graphcpp;
 
-Edge::Edge(unsigned int v1, unsigned int v2) noexcept:
+Edge::Edge(ushort v1, ushort v2) noexcept:
 	_v1(std::min(v1, v2)), _v2(std::max(v1, v2)) 
 {
 	assert(v1 != v2);
@@ -21,11 +21,11 @@ bool Edge::operator!=(const Edge& right) const noexcept
 	return !(*this == right);
 }
 
-unsigned int Edge::v1() const noexcept
+ushort Edge::v1() const noexcept
 {
 	return _v1;
 }
-unsigned int Edge::v2() const noexcept
+ushort Edge::v2() const noexcept
 {
 	return _v2;
 }
